rustdoc + cleanup fuel-indexer-lib
