+++ packages/fuel-indexer-tests/src/fixtures.rs
@@ -5,7 +5,7 @@ use fuel_indexer_database::IndexerConnectionPool;
 use fuel_indexer_lib::{
     config::{
         auth::AuthenticationStrategy, defaults as config_defaults, AuthenticationConfig,
-        DatabaseConfig, FuelNodeConfig, GraphQLConfig, IndexerConfig, RateLimitConfig,
+        DatabaseConfig, FuelClientConfig, GraphQLConfig, IndexerConfig, RateLimitConfig,
     },
     defaults::SERVICE_REQUEST_CHANNEL_SIZE,
     utils::{derive_socket_addr, ServiceRequest},
@@ -287,7 +287,7 @@ pub async fn api_server_app_postgres(
         verbose: true,
         local_fuel_node: false,
         indexer_net_config: false,
-        fuel_node: FuelNodeConfig::default(),
+        fuel_node: FuelClientConfig::default(),
         database,
         graphql_api: GraphQLConfig::default(),
         metrics: false,
@@ -322,7 +322,7 @@ pub async fn authenticated_api_server_app_postgres(database_url: Option<&str>) -
         verbose: true,
         local_fuel_node: false,
         indexer_net_config: false,
-        fuel_node: FuelNodeConfig::default(),
+        fuel_node: FuelClientConfig::default(),
         database,
         graphql_api: GraphQLConfig::default(),
         metrics: false,
@@ -363,7 +363,7 @@ pub async fn indexer_service_postgres(
         verbose: true,
         local_fuel_node: false,
         indexer_net_config: false,
-        fuel_node: FuelNodeConfig::default(),
+        fuel_node: FuelClientConfig::default(),
         database,
         graphql_api: GraphQLConfig::default(),
         metrics: false,
