+++ plugins/forc-index/src/commands/deploy.rs
@@ -1,4 +1,4 @@
-use crate::{ops::forc_index_deploy, utils::defaults};
+use crate::{defaults, ops::forc_index_deploy};
 use anyhow::Result;
 use clap::Parser;
 use std::path::PathBuf;
