+++ plugins/forc-index/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod cli;
 pub mod commands;
+pub(crate) mod defaults;
 pub(crate) mod ops;
 pub(crate) mod utils;
